version 0.93

+++ src/poly_mult.c
@@ -1,4 +1,4 @@
-void poly_mult(double *a, long *na, double *b, long *nb, double *ab) {
+void poly_mult(double *a, int *na, double *b, int *nb, double *ab) {
   int i, j, nab = *na + *nb - 1;
 
   for(i = 0; i < nab; i++) ab[i] = 0.0;
